requirements.txt
numpy
scipy
tqdm
pytest-runner
librosa
python-Levenshtein
visualdl
SoundFile
resampy
webrtcvad
paddlepaddle